тест

+++ notes.controller.js
@@ -1,9 +1,9 @@
 const chalk = require('chalk');
 const Note = require('./models/Note');
 
-async function addNote(title) {
-    await Note.create({ title });
-    console.log(chalk.bgGreen('Note added'));
+async function addNote(title, owner) {
+    await Note.create({ title, owner });
+    console.log(chalk.bgGreen('Заметка добавлена'));
 }
 
 async function getNotes() {
@@ -11,14 +11,20 @@ async function getNotes() {
     return notes;
 }
 
-async function removeNote(id) {
-    await Note.deleteOne({ _id: id });
-    console.log(chalk.bgRed('Note removed'));
+async function removeNote(id, owner) {
+    const result = await Note.deleteOne({ _id: id, owner });
+    if (result.deletedCount === 0) {
+        throw new Error('Заметка не найдена');
+    }
+    console.log(chalk.bgRed('Заметка удалена'));
 }
 
-async function updateNote(id, title) {
-    await Note.updateOne({ _id: id }, { title });
-    console.log(chalk.bgYellow('Note updated'));
+async function updateNote(id, title, owner) {
+    const result = await Note.updateOne({ _id: id, owner }, { title });
+    if (result.matchedCount === 0) {
+        throw new Error('Заметка не найдена');
+    }
+    console.log(chalk.bgYellow('Заметка обновлена'));
 }
 
 module.exports = {
